prefactor solrman smmodel code (#13607)
