+++ solrman/smmodel/model.go
@@ -15,7 +15,6 @@
 package smmodel
 
 import (
-	"encoding/json"
 	"fmt"
 	"sort"
 	"sync"
@@ -276,157 +275,4 @@ func (m *Model) ComputeBestMoves(numCPU int, count int) []Move {
 		}
 	}
 	return moves
-}
-
-type Node struct {
-	Name       string  `json:"name"`
-	Address    string  `json:"address"`
-	Evacuating bool    `json:"evacuating"`
-	Docs       float64 `json:"docs"`
-	Size       float64 `json:"size"` // in bytes
-	coreCount  int
-	id         int // unique id of this node for internal tracking
-
-	cost float64 // cached cost
-}
-
-func (n *Node) Add(core *Core) {
-	n.Docs += core.Docs
-	n.Size += core.Size
-	n.coreCount += 1
-	core.nodeId = n.id
-}
-
-func (n *Node) Contains(core *Core) bool {
-	return core.nodeId == n.id
-}
-
-func (n *Node) With(core *Core) *Node {
-	if !n.Contains(core) {
-		panic(fmt.Sprintf("core %s assigned to %d cannot be added to %d", core.Name, core.nodeId, n.id))
-	}
-
-	nCopy := *n
-	nCopy.cost = 0
-
-	nCopy.Docs += core.Docs
-	nCopy.Size += core.Size
-	nCopy.coreCount++
-	return &nCopy
-}
-
-func (n *Node) Without(core *Core) *Node {
-	if !n.Contains(core) {
-		return n
-	}
-
-	nCopy := *n
-	nCopy.cost = 0
-
-	nCopy.Docs -= core.Docs
-	nCopy.Size -= core.Size
-	nCopy.coreCount--
-	return &nCopy
-}
-
-type Collection struct {
-	Name  string  `json:"name"`
-	Docs  float64 `json:"docs"`
-	Size  float64 `json:"size"` // in bytes
-	cores []*Core
-
-	cost float64 // cached cost
-}
-
-func (c *Collection) Add(core *Core) {
-	c.Docs += core.Docs
-	c.Size += core.Size
-	c.cores = append(c.cores, core)
-}
-
-func (c *Collection) Without(core *Core) *Collection {
-	found := false
-	newCores := make([]*Core, 0, len(c.cores))
-	for _, c := range c.cores {
-		if c == core {
-			found = true
-		} else {
-			newCores = append(newCores, c)
-		}
-	}
-
-	if !found {
-		panic(fmt.Sprintf("core %s not found in collection %s", core.Name, c.Name))
-	}
-
-	return &Collection{
-		Name:  c.Name,
-		Docs:  c.Docs - core.Docs,
-		Size:  c.Size - core.Size,
-		cores: newCores,
-	}
-}
-
-func (c *Collection) Replace(core *Core, newCore *Core) *Collection {
-	found := false
-	newCores := make([]*Core, 0, len(c.cores))
-	for _, c := range c.cores {
-		if c == core {
-			found = true
-			newCores = append(newCores, newCore)
-		} else {
-			newCores = append(newCores, c)
-		}
-	}
-
-	if !found {
-		panic(fmt.Sprintf("core %s not found in collection %s", core.Name, c.Name))
-	}
-
-	return &Collection{
-		Name:  c.Name,
-		Docs:  c.Docs - core.Docs + newCore.Docs,
-		Size:  c.Size - core.Size + newCore.Size,
-		cores: newCores,
-	}
-}
-
-type Core struct {
-	Name       string  `json:"name"`
-	Collection string  `json:"collection"`
-	Shard      string  `json:"shard"`
-	Docs       float64 `json:"docs"`
-	Size       float64 `json:"size"` // in bytes
-	nodeId     int     // the node I currently belong to
-}
-
-func (c *Core) shardName() string {
-	return c.Collection + "_" + c.Shard
-}
-
-type Move struct {
-	Core     *Core
-	FromNode *Node
-	ToNode   *Node
-}
-
-func (m *Move) String() string {
-	result, _ := json.Marshal(m.AsJson())
-	return string(result)
-}
-
-func (m *Move) AsJson() interface{} {
-	return &struct {
-		Name       string `json:"core"`
-		Collection string `json:"collection"`
-		Shard      string `json:"shard"`
-		FromNode   string `json:"from_node"`
-		ToNode     string `json:"to_node"`
-	}{
-		Name:       m.Core.Name,
-		Collection: m.Core.Collection,
-		Shard:      m.Core.Shard,
-		FromNode:   m.FromNode.Name,
-		ToNode:     m.ToNode.Name,
-	}
 }
