+++ solrman/smmodel/move.go
@@ -1,0 +1,44 @@
+// Copyright 2019 FullStory, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package smmodel
+
+import "encoding/json"
+
+type Move struct {
+	Core     *Core
+	FromNode *Node
+	ToNode   *Node
+}
+
+func (m *Move) String() string {
+	result, _ := json.Marshal(m.AsJson())
+	return string(result)
+}
+
+func (m *Move) AsJson() interface{} {
+	return &struct {
+		Name       string `json:"core"`
+		Collection string `json:"collection"`
+		Shard      string `json:"shard"`
+		FromNode   string `json:"from_node"`
+		ToNode     string `json:"to_node"`
+	}{
+		Name:       m.Core.Name,
+		Collection: m.Core.Collection,
+		Shard:      m.Core.Shard,
+		FromNode:   m.FromNode.Name,
+		ToNode:     m.ToNode.Name,
+	}
+}
